Set proxy and proxy settings for Telegram bot because of blocking t.me in Russia.
Use list of proxies as settings in Django app (need to use api for getting proxies)
As a cheking system it pings t.me site and if it isn't reachable it pings throught proxies until it has available proxy.
Then it resets proxy for bot. And if raised MaxRetryError for messages it resets proxy.