Cover whole ResaleListView with tests and fix bug in list filtering
Create 18 tests for ResaleListView including next filters:
- default list test - 4 tests
- price (from, to) - 4 tests
- area (from, to) - 4 tests
- rooms - 2 tests
- neighbourhood - 2 tests
- any_text - 1 test
All of them combine test with and without is_active and status.

Fix bug in
real_estate.views.ApartmentFilterMixin.filterApartmentCheckbox
wich is parent for both ResaleListView and ResidentalComplexList

Bug was related to gte filter in queryset.
It checked for rooms_gte=4, for 4+ checkbox, but 'B'>4 where
'B' is Bachelor.